lint.
git-svn-id: file:///svn/unbound/trunk@4022 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/shm_side/shm_main.c
@@ -84,7 +84,7 @@ int shm_main_init(struct daemon* daemon)
 {
 #ifdef HAVE_SHMGET
 	struct shm_stat_info *shm_stat;
-	int shm_size;
+	size_t shm_size;
 	
 	/* sanitize */
 	if(!daemon)
